# docs/swagger.yaml
basePath: /
definitions:
  handlers.NoContentResponse:
    type: object
  handlers.ProfileResponse:
    properties:
      created_at:
        description: user create datetime
        example: "0001-01-01T00:00:00Z"
        type: string
      email:
        description: email
        example: some@mail.com
        type: string
      login:
        description: login
        example: some-login
        type: string
      name:
        description: name
        example: some-name
        type: string
      phone:
        description: phone
        example: "88009998889988"
        type: string
    type: object
  handlers.UserAuthorizationArgs:
    properties:
      login:
        description: login
        example: some login
        type: string
      password:
        description: password
        example: some_password
        type: string
    type: object
  handlers.UserAuthorizationResponse:
    properties:
      expired_at:
        description: token deadline date time
        type: integer
      token:
        description: access token string
        type: string
    type: object
  handlers.UserCreateArgs:
    properties:
      email:
        description: email
        example: some@mail.com
        maxLength: 32
        minLength: 5
        type: string
      login:
        description: login
        example: somelogin
        type: string
      name:
        description: name
        example: some name
        type: string
      password:
        description: password
        example: secret-word
        maxLength: 1024
        minLength: 7
        type: string
      phone:
        description: phone
        example: "88009998889999"
        type: string
    type: object
  handlers.UserPasswordEditArgs:
    properties:
      password:
        description: new password
        example: some_password
        type: string
    type: object
  liberror.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is a growth service for managing personal targets
  license:
    name: MIT
  title: Growth-place API
  version: "1.0"
paths:
  /user:
    delete:
      consumes:
      - application/json
      description: mark user as deleted
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.NoContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/liberror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/liberror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/liberror.Error'
      summary: delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: returns user personal data
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/liberror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/liberror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/liberror.Error'
      summary: get profile
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create new user on system
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UserCreateArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.NoContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/liberror.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/liberror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/liberror.Error'
      summary: new user
      tags:
      - users
  /user/authorization:
    post:
      consumes:
      - application/json
      description: authorization user on system by email and password
      parameters:
      - description: User auth data
        in: body
        name: authorization
        required: true
        schema:
          $ref: '#/definitions/handlers.UserAuthorizationArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserAuthorizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/liberror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/liberror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/liberror.Error'
      summary: authorization
      tags:
      - users
  /user/password:
    post:
      consumes:
      - application/json
      description: 'edit user password: add new or replace old password'
      parameters:
      - description: New password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/handlers.UserPasswordEditArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.NoContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/liberror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/liberror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/liberror.Error'
      summary: edit password
      tags:
      - users
swagger: "2.0"